requirements.txt
anthropic
gspread
google-auth
google-auth-oauthlib
requests
python-telegram-bot
pdfplumber
pytz
playwright
python-dotenv
beautifulsoup4
lxml
openai
google-generativeai